refactor(patchers): get rid of @nextcloud/initial-state patching
